Allow multiple processes to work together
Using shared memory, multiple processes can now use persistent
threads.

Issues to be resolved:
1) Place std::map on shared memory using boost::interprocess lib.
2) Find the exact reason for latency overhead in persistent model
as compared to native code. Check if this overhead is fixed or
per iteration (We have narrowed it to the delay introduced in
launching kernel by waiting).
3) Synchronization needs to be completed. Care must be taken
for keeping all data of primitives on shared memory and not on
heap.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>