Device: Static/Dynamic queuing hybrid approach
Currently, allocation of blocks to SM blocks is pure dymanic
using atomicAdd(). But this can cause performance issues, specially when
cuda kernel is so small that majority of time is spent in atomicAdd(). To
solve this problem, a hybrid queuing approach is used (has to be specially
invoked by programmer)

Signed-off-by: Saksham Jain <sakjain92@gmail.com>