Simplify userspace coloring code in kernel driver
Coloring inside kernel can be broken down into two parts:
1) Allocation:
For userspace coloring, only condition neccesary is contiguous allocation
of physical memory.
2) Memcpy/Transfer:
This needs to be color aware inside kernel. The only reason that this needs
to be done inside kernel is because of performance reasons.

TODO:
Memcpy in kernel needs to be implemented.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>