Implement colored memcpy in kernel
For userspace coloring, because the colored pages are not physically/virtually
contiguous, traditional memcpy will not work. Presently the memcpy was
implemented in userspace (by skipping pages of different colors). This
was possible because userspace knows about the physical addresses.

But this approach was significantly slower because independent calls to
cudaMemcpy was most likely not pipelined inside kernel. Hence it was
needed to implement a kernel function to achieve the same.

Note: Present kernel function is very limited. It requires all the pages
(on GPU and CPU) to be populated prior to calling it (Otherwise it will
fail and not crash, thankfully). Also, it requires the CPU side pages to
be allocated using cudaMallocManaged().

TODO: Need to similarly implement cudaMemset()
TODO: Remove the need for pages to be pre-populated
TODO: Remove the need for CPU memory to be allocated using
cudaMallocManaged()

Signed-off-by: Saksham Jain <sakjain92@gmail.com>