Add colored memset
Similarly to memcpy, for userspace coloring, kernel supported memset is
needed.

Note: Colored Memset is quite slower than cudaMemset()
This can be explained by probably three reasons:

1) cudaMemset() seems to be asynchronous to the CPU. So it doesn't wait
for GPU operation to complete before returning to CPU whereas currently
our colored memset is synchronous.

2) Colored memset is currently using GPU function memset_1() which sets
memory 1 byte at a time. This can be improved by using memset_8() but
alignment issues need to be considered

3) For cudaMemset(), all GPU memory is contiguous (so only a single
command need to be issued to the DMA) but colored memory is fragmented
for the most part.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>