Speedup colored memcpy a bit
By consolidating physically contiguous pages we can speedup colored memcpy by
issuing less number of total DMA operations.

It seems that for colored memory the bottleneck seems to be the CPU
operations since for userspace colored memory, we need to iterate
over all physical address of pages to find which one of them is of the
right color whereas the traditional cudaMemcpy can just issue one single
DMA request because the memory is contiguous on GPU mostly.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>