Add paranoid check flag
There is a check we do currently that is not necessary but adds to
delays/latency. We have added an optional flag (disabled by default) for
this check.

The check: Our approach to SM affinity is to launch more blocks than necessary
to populate the whole GPU and then allow blocks that are running on
wrong SMs to exit. For this approach to work, we have to assume that
hardware schedules blocks at much faster rate than execution of blocks
so blocks don't exit before all blocks are scheduled. We can avoid this
assumption using a software check, but this check seems to be
unnecessary so can be skipped.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>