Use cuda function to find out number of persistent blocks
The total number of resident blocks in an SM depends on:

1) A hard limit on number of blocks per SM.
2) Number of threads per block.
3) Amount of shared memory per block.
4) Number of registers used per block.

Since it is difficult to probe 3) and 4) without cuda support,
we use cuda provided function (cudaOccupancyMaxActiveBlocksPerMultiprocessor)
to find out number of resident blocks.

Earlier we were assuming that only 2) mattered (This assumption holds true
for most of the code but not all) hence we could do without using cuda
provided function.

cudaOccupancyMaxActiveBlocksPerMultiprocessor() seems to be lightweight
function (< 1 us).

Signed-off-by: Saksham Jain <sakjain92@gmail.com>