Specify exact CUDA function when function is overloaded
Due to change in FGPU API implementation, we require to know which
eaxct function is being used. This is a problem when function is overloaded.
For now, we pass on the responsibility to the user.

TODO: Look at
"https://stackoverflow.com/questions/9054703/overloaded-function-as-argument-of-variadic-template-function"
It might have solution as how to use C++ advance features to make it
possible to hide this side-effect from user.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>