Implement faster inter-block synchronization on GPU
For compute coloring, it is required that p-blocks only exit after
all p-blocks have been launched. To do this, a inter-block sychronization
is required. For this, earlier we were using a lockless implementation.
But the number of reads required in this approach as quite large. Hence
it performed slow (We could have increased performance by reading
all values in parallel, but because we don't know how many threads in
a block, we are forced to do it serially). To improve performance,
we use a lock-based implementation. According to micro-benchmarks,
this should work on sub-microsecond speed.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>