Bugfix: Find the lowest FD which connects with MPS
We are trying to find the lowest FD which connects with MPS. Hence
once we find it, we store it in g_uvm_fd. We shouldn't set this
variable again in future.

TODO: A better approach is to find the FD using /proc/<pid>/fd.
Find the file nvidia-uvm and find the corresponding FD. This way we
don't need to overload open() and connect() functions. So we don't
require LD_PRELOAD.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>