Modify server to run in background
When server is launched as background process, stdin would be closed
by bash and server, which is waiting for input on stdin, would close.
So to remedy this, whenever stdin is closed, the process is sent int
wait mode, waiting for a signal.

Also the server binary has been renamed to fgpu_server to make it easier
to find and kill the server in a script.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>