FGPU has internal memory overheads
As we have started coloring the metadata also, FGPU has some overhead
when it comes to memory allocation (as it internally uses some memory).
Reverse engineering code currently tries to greedily consume all available
memory. But this is not possible since some memory is already consumed by the
time reverse engineering code is launched. So as a temporary hack, reverse
engineering code assumes so constant overhead and take it as margin.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>